# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação API controle de acesso Laravel",
        "description": "Documentação com todos os endpoints necessários para a correta utilização da aplicação.",
        "contact": {
            "email": "assuncaofilipe97@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/app/user/register": {
            "post": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Cadastro de usuário para utilização da API",
                "description": "Retorna as informações do usuário cadastrado",
                "operationId": "00d85c1350d1b1825d8a8ecc0c48a99e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "email_confirmation": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "terms_of_use": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "register": {
                                    "summary": "Cadastro de usuário",
                                    "value": {
                                        "email": "exemplo@gmail.com",
                                        "email_confirmation": "exemplo@gmail.com",
                                        "password": "123a123a",
                                        "password_confirmation": "123a123a",
                                        "terms_of_use": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cadastro feito com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/loginResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "email": {
                                                    "type": "string"
                                                },
                                                "password": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "obrigatório": {
                                        "summary": "Campos obrigatórios",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "email": [
                                                    "O campo email é obrigatório."
                                                ],
                                                "password": [
                                                    "O campo senha é obrigatório."
                                                ],
                                                "terms_of_use": [
                                                    "O campo termos de uso é obrigatório."
                                                ]
                                            }
                                        }
                                    },
                                    "comfirmação": {
                                        "summary": "Campos de confirmação / Regras inválidas",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "email": [
                                                    "O campo email deve ser um endereço de e-mail válido.",
                                                    "O campo email de confirmação não confere."
                                                ],
                                                "password": [
                                                    "O campo senha de confirmação não confere.",
                                                    "O campo senha deve ter pelo menos 8 caracteres.",
                                                    "O campo senha deve conter pelo menos uma letra."
                                                ],
                                                "terms_of_use": [
                                                    "É obrigatório o aceite dos Termos de uso"
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "loginResponse": {
                "properties": {
                    "success": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "access_token": {
                                "type": "string"
                            },
                            "token_type": {
                                "type": "string"
                            },
                            "expires_in": {
                                "type": "string",
                                "format": "date"
                            },
                            "user": {
                                "$ref": "#/components/schemas/user"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "success": "true",
                    "data": {
                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC90cm91dy1hcHBcL2FwcFwvYXV0aFwvbG9naW4iLCJpYXQiOjE2NDQzMjYwODksImV4cCI6MTY0NDMyOTY4OSwibmJmIjoxNjQ0MzI2MDg5LCJqdGkiOiJDeENTS0x6cGRLQ3ZUc2ZQIiwic3ViIjoxOTUsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.D8i4qR2vgW9G-lz-ej1swDUToFgU4AnKHEFY_2hWRGA",
                        "token_type": "bearer",
                        "expires_in": 3600,
                        "user": {
                            "id": 195,
                            "uuid": "d83810de-4d4a-4306-85ed-c2e761fc998a",
                            "email": "teste2@gmail.com",
                            "email_verified_at": null,
                            "created_at": "2022-02-08T13:09:57.000000Z",
                            "updated_at": "2022-02-08T13:09:57.000000Z"
                        }
                    }
                }
            },
...
            "user": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "email_verified_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}